+++ app/controllers/staff/planned_disbursement_uploads_controller.rb
@@ -6,26 +6,30 @@ class Staff::PlannedDisbursementUploadsController < Staff::BaseController
   include Secured
   include StreamCsvDownload
 
-  before_action :authorize_report
-
   def new
-    @report_presenter = ReportPresenter.new(@report)
+    authorize report, :show?
+
+    @report_presenter = ReportPresenter.new(report)
   end
 
   def show
-    @report_presenter = ReportPresenter.new(@report)
-    generator = ImportPlannedDisbursements::Generator.new(@report)
+    authorize report, :show?
+
+    @report_presenter = ReportPresenter.new(report)
+    generator = ImportPlannedDisbursements::Generator.new(report)
     filename = @report_presenter.filename_for_forecasts_template
 
     stream_csv_download(filename: filename, headers: generator.column_headings) do |csv|
-      @report.reportable_activities.each do |activity|
+      report.reportable_activities.each do |activity|
         csv << generator.csv_row(activity)
       end
     end
   end
 
   def update
-    @report_presenter = ReportPresenter.new(@report)
+    authorize report, :upload?
+
+    @report_presenter = ReportPresenter.new(report)
     upload = CsvFileUpload.new(params[:report], :planned_disbursement_csv)
     @success = false
 
@@ -44,8 +48,7 @@ class Staff::PlannedDisbursementUploadsController < Staff::BaseController
     end
   end
 
-  private def authorize_report
-    @report = Report.find(params[:report_id])
-    authorize @report, :show?
+  private def report
+    @_report ||= Report.find(params[:report_id])
   end
 end