+++ CHANGELOG.md
@@ -646,6 +646,8 @@
 - Update the request an account link
 - Allow DP users to create transfers and associate them with the current report
 - Add some useful stats to our healthcheck endpoint
+- Prevent uploading transactions if the report is not editable
+- Prevent uploading activities, transactions and planned disbursements if the report is not editable
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-50...HEAD
 [release-50]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-49...release-50